refactor(optimize): move topology cleanup and setup from iteration.rs to run_loop.rs
- iteration.rs now contains only branch-length bookkeeping (save, restore, damping)
- Topology cleanup (find_zero_optimal_internal_edges, prune_and_merge_in_loop) and setup helpers (apply_initial_guess_mode, normalize_partition_rates, branch-length validators) move to run_loop.rs, their only production caller
