+++ packages/treetime/src/optimize/__tests__/test_dispatch_zero_boundary.rs
@@ -4,7 +4,7 @@ mod tests {
   use crate::ancestral::marginal::{initialize_marginal, update_marginal};
   use crate::gtr::get_gtr::{GtrModelName, JC69Params, get_gtr_by_name, jc69};
   use crate::optimize::params::BranchOptMethod;
-  use crate::optimize::iteration::find_zero_optimal_internal_edges;
+  use crate::optimize::run_loop::find_zero_optimal_internal_edges;
   use crate::optimize::method_newton::{newton_inner, newton_sqrt_inner};
   use crate::optimize::dispatch::{initial_guess_mixed, run_optimize_mixed};
   use crate::optimize::likelihood::{evaluate_mixed, evaluate_mixed_log_lh_only};
