+++ packages/treetime/src/optimize/__tests__/test_initial_guess_mode.rs
@@ -4,7 +4,7 @@ mod tests {
   use crate::ancestral::marginal::{initialize_marginal, update_marginal};
   use crate::gtr::get_gtr::{JC69Params, jc69};
   use crate::optimize::params::InitialGuessMode;
-  use crate::optimize::iteration::{
+  use crate::optimize::run_loop::{
     any_edge_missing_branch_length, any_indel_edge_has_zero_branch_length, apply_initial_guess_mode,
   };
   use crate::optimize::dispatch::initial_guess_mixed;
