+++ packages/treetime/src/optimize/dispatch.rs
@@ -6,7 +6,7 @@ use crate::optimize::params::BranchOptMethod;
 use crate::optimize::zero_boundary::{is_zero_branch_optimal, min_branch_length_for_indels, reconcile_zero_boundary};
 use crate::representation::partition::optimization_contribution::OptimizationContribution;
 use crate::representation::partition::traits::PartitionOptimizeOps;
-use crate::{make_error, make_internal_report};
+use crate::make_error;
 use eyre::Report;
 use parking_lot::RwLock;
 use std::sync::Arc;
